Cleaner implementation, share utility methods in 'Tokens' class
